# week-4/BE-03/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BE-03 — Supabase Auth API",
    "version": "1.0.0",
    "description": "A demo Express API showcasing authentication flows with Supabase: signup, login, protected routes, and logout."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/auth/signup": {
      "post": {
        "summary": "Create a new user account",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 6 }
                }
              },
              "example": {
                "email": "user@example.com",
                "password": "securePass123"
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User created successfully" },
          "400": { "description": "Missing or invalid fields" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Authenticate and receive JWT tokens",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" }
                }
              },
              "example": {
                "email": "user@example.com",
                "password": "securePass123"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": { "type": "string" },
                    "refresh_token": { "type": "string" },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string" },
                        "email": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Invalid login credentials" }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "End the current session (requires Bearer token)",
        "tags": ["Auth"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "204": { "description": "Logged out successfully — no content" },
          "401": { "description": "Invalid or expired token" }
        }
      }
    },
    "/public/info": {
      "get": {
        "summary": "Public endpoint — no auth required",
        "tags": ["Public"],
        "responses": {
          "200": {
            "description": "Returns a welcome message",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/protected/profile": {
      "get": {
        "summary": "Get authenticated user profile",
        "tags": ["Protected"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "User profile data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string" },
                    "email": { "type": "string" },
                    "created_at": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": { "description": "Invalid or expired token" }
        }
      }
    },
    "/protected/dashboard": {
      "get": {
        "summary": "Access the protected dashboard",
        "tags": ["Protected"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Dashboard welcome message with user info",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "user_id": { "type": "string" },
                    "email": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": { "description": "Invalid or expired token" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Paste your Supabase JWT access token"
      }
    }
  }
}